Making an RSVP Application with JS
Using event handlers to make an interactive RSVP list. 

We started by selecting our form and input elements by creating global constants. We then used an event handler and callback function to create a new list element (li) within our ul. We then set its text content to that of the input field in the HTML file, resetting the text content of the input to an empty field. Along with the user input textContent, we added children in a label (and a checkbox). 

We then created another event handler for responding to the checkbox, this listened for 'change'. Thus we created a const checkbox for the target, a const checked for whenever it was checked and a const listItem for the grandparent list item defined earlier. We then wrote an if statement that stated if the checkbox was checked, it would change the list item class name to 'responded'. Else the class name would be an empty string.

We then looked at our code, understanding the advantages of modules. As a way to clean our code and make it easier to read, we went back to our form event handler and extracted all of the code around element creation and created a preceeding function, 'createLi'. As suggested, createLi would create the list element (and the previously defined children) with the parameter of 'text'. eg. the user input. We then added a return for the contents of the function as JS does not by default and proceeded to add our function back into the event handler for the form.